use closure instead of userdata with metatable for js functions
Using userdata with a metatable is a cool idea but it's kind of fragile.
Because it's userdata, its type isn't function and therefore it cannot
be used to do things like package.preload values. It also means that it
could be overwritten during runtime, which would be weird. This works
much better.